Manual sync with global requirements
shade, oaktreemodel and grpc are not in global-requirements, so this has
to be done by hand / with care. We need to remove python-*client from
shade before we can put it in the global-requirements, otherwise the
client lib versions will cause epic amounts of pain. So until we do that
we really can't apply for official status or directly follow
requirements. However, the intent is to follow them, so this will just
have to be something we keep up with by hand a bit.

Change-Id: Ie9fd8f121686f84b32f50371b24a247f86421b26