Add zuul v3 job to trigger readthedocs
Add a .zuul.yaml and consume the docs-on-readthedocs project-template.
Also go ahead and consume openstack-python-jobs, openstack-python35-jobs
and publish-to-pypi-quietly templates to verify that they work.

Add an entry to the TODO list to make better jobs, which isn't super
important other than as a way to see that the trigger worked.

Change-Id: I954f8938ac6d0b97bdd9cf6b9e436423f4cee6f3
Depends-On: I75850f8f6e3a4b7eb796afa4f5ea947dfce347f8